Create hybridctl Delete CMD - deployment

+++ hcp-apiserver/pkg/handler/hcpResourceHandler.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"Hybrid_Cloud/hybridctl/util"
 	cobrautil "Hybrid_Cloud/hybridctl/util"
 	resourcev1alpha1 "Hybrid_Cloud/pkg/apis/resource/v1alpha1"
 	resourcev1alpha1clientset "Hybrid_Cloud/pkg/client/resource/v1alpha1/clientset/versioned"
@@ -10,6 +11,7 @@ import (
 	"io/ioutil"
 	"net/http"
 
+	"github.com/gorilla/mux"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -24,7 +26,6 @@ type Resource struct {
 func CreateDeploymentHandler(w http.ResponseWriter, r *http.Request) {
 
 	var resource Resource
-
 	jsonDataFromHttp, _ := ioutil.ReadAll(r.Body)
 	err := json.Unmarshal(jsonDataFromHttp, &resource)
 	if err != nil {
@@ -39,6 +40,18 @@ func CreateDeploymentHandler(w http.ResponseWriter, r *http.Request) {
 	bytes, _ := json.Marshal(resource.RealResource)
 	json.Unmarshal(bytes, &real_resource)
 
+	master_config, err := cobrautil.BuildConfigFromFlags("kube-master", "/root/.kube/config")
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	master_clienset, err := resourcev1alpha1clientset.NewForConfig(master_config)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
 	/*
 		// HCPPolicy 최적 배치 알고리즘 정책 읽어오기
 		algorithm, err := policy.GetAlgorithm()
@@ -50,41 +63,12 @@ func CreateDeploymentHandler(w http.ResponseWriter, r *http.Request) {
 	*/
 
 	// TargetCluster가 지정되지 않은 경우
-	if resource.TargetCluster == "undefined" {
-
-		master_config, err := cobrautil.BuildConfigFromFlags("kube-master", "/root/.kube/config")
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		clienset, err := resourcev1alpha1clientset.NewForConfig(master_config)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
+	if resource.TargetCluster == "" {
 
 		// HCPDeployment 생성하기
-		hcp_resource := resourcev1alpha1.HCPDeployment{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "HCPDeployment",
-				APIVersion: "hcp.crd.com/v1alpha1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name: real_resource.Name,
-			},
-			Spec: resourcev1alpha1.HCPDeploymentSpec{
-				RealDeploymentSpec:     real_resource.Spec,
-				RealDeploymentMetadata: real_resource.ObjectMeta,
+		hcp_resource := deploymentToHCPDeployment(real_resource)
 
-				// SchedulingStatus "Requested"
-				SchedulingNeed:     true,
-				SchedulingComplete: false,
-				//SchedulingType:   algorithm[0],
-			},
-		}
-
-		r, err := clienset.HcpV1alpha1().HCPDeployments("hcp").Create(context.TODO(), &hcp_resource, metav1.CreateOptions{})
+		r, err := master_clienset.HcpV1alpha1().HCPDeployments("hcp").Create(context.TODO(), &hcp_resource, metav1.CreateOptions{})
 		if err != nil {
 			fmt.Println(err)
 			return
@@ -105,6 +89,18 @@ func CreateDeploymentHandler(w http.ResponseWriter, r *http.Request) {
 			namespace = "default"
 		}
 
+		hcp_resource := deploymentToHCPDeployment(real_resource)
+		hcp_resource.Spec.SchedulingResult.Targets = append(hcp_resource.Spec.SchedulingResult.Targets, resourcev1alpha1.Target{
+			Cluster:  resource.TargetCluster,
+			Replicas: real_resource.Spec.Replicas,
+		})
+		_, err = master_clienset.HcpV1alpha1().HCPDeployments("hcp").Create(context.TODO(), &hcp_resource, metav1.CreateOptions{})
+		if err != nil {
+			fmt.Println(err)
+			return
+		} else {
+			fmt.Printf("succeed to create hcpdeployment: %s \n", hcp_resource.Name)
+		}
 		// Kubernetes Deployment 생성
 		r, err := clientset.AppsV1().Deployments(namespace).Create(context.TODO(), real_resource, metav1.CreateOptions{})
 
@@ -118,24 +114,42 @@ func CreateDeploymentHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func DeleteDeploymentHandler(w http.ResponseWriter, r *http.Request) {
-	vals := r.URL.Query()
-	target_cluster := vals.Get("cluster")
-	namespace := vals.Get("namespace")
-	name := vals.Get("name")
-	config, err := cobrautil.BuildConfigFromFlags(target_cluster, "/root/.kube/config")
-	if err != nil {
-		fmt.Println(err)
-	}
-	clientset, _ := kubernetes.NewForConfig(config)
+	vars := mux.Vars(r)
+	namespace := vars["namespace"]
+	name := vars["name"]
+	master_config, _ := cobrautil.BuildConfigFromFlags("kube-master", "/root/.kube/config")
+	master_clientset, _ := resourcev1alpha1clientset.NewForConfig(master_config)
 
-	// create resource
-	err = clientset.AppsV1().Deployments(namespace).Delete(context.TODO(), name, metav1.DeleteOptions{})
+	hcpdeployment, err := master_clientset.HcpV1alpha1().HCPDeployments("hcp").Get(context.TODO(), name, metav1.GetOptions{})
+
+	if !hcpdeployment.Spec.SchedulingNeed && hcpdeployment.Spec.SchedulingComplete {
+		// if target_cluster != "" {
+		// 	config, _ := cobrautil.BuildConfigFromFlags(target_cluster, "/root/.kube/config")
+		// 	clientset, _ := kubernetes.NewForConfig(config)
+		// 	err = clientset.AppsV1().Deployments(namespace).Delete(context.TODO(), name, metav1.DeleteOptions{})
+		// } else {
+		targets := hcpdeployment.Spec.SchedulingResult.Targets
+		for _, target := range targets {
+			// TODO : cluster unregister한 경우
+			config, _ := util.BuildConfigFromFlags(target.Cluster, "/root/.kube/config")
+			clientset, _ := kubernetes.NewForConfig(config)
+			err = clientset.AppsV1().Deployments(namespace).Delete(context.TODO(), name, metav1.DeleteOptions{})
+		}
+		// }
+	}
 
 	if err != nil {
 		fmt.Println(err)
 		return
 	} else {
-		fmt.Printf("success to delete deployment %s \n", name)
+		fmt.Printf("succeed to delete deployment %s \n", name)
+		err = master_clientset.HcpV1alpha1().HCPDeployments("hcp").Delete(context.TODO(), name, metav1.DeleteOptions{})
+		if err != nil {
+			fmt.Println(err)
+			return
+		} else {
+			fmt.Printf("succeed to delete hcpdeployment %s \n", name)
+		}
 	}
 }
 
@@ -232,4 +246,27 @@ func CreatePodHandler(w http.ResponseWriter, r *http.Request) {
 			fmt.Printf("success to create pod %s \n", r.Name)
 		}
 	}
+}
+
+func deploymentToHCPDeployment(real_resource *appsv1.Deployment) resourcev1alpha1.HCPDeployment {
+	// HCPDeployment 생성하기
+	hcp_resource := resourcev1alpha1.HCPDeployment{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "HCPDeployment",
+			APIVersion: "hcp.crd.com/v1alpha1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: real_resource.Name,
+		},
+		Spec: resourcev1alpha1.HCPDeploymentSpec{
+			RealDeploymentSpec:     real_resource.Spec,
+			RealDeploymentMetadata: real_resource.ObjectMeta,
+
+			// SchedulingStatus "Requested"
+			SchedulingNeed:     true,
+			SchedulingComplete: false,
+			//SchedulingType:   algorithm[0],
+		},
+	}
+	return hcp_resource
 }
